+++ src/index.ts
@@ -6,3 +6,4 @@ export * from './sync';
 export * from './install';
 export * from './update';
 export * from './check';
+export * from './uninstall';